Crosshair Texture Rectangle Fix & Readme Updates
For some reason the crosshair was fighting me once I refactored the weapon manager. Thus I've created the crosshair textureRect as part of the camera instead of by code, and proceed as normal.

Not entirely sure if the scaling and such is the best.

Credit to Agnis NeZvers for video providing useful info.

Updating Readme in prep for future commits.