Class Names, Health Bar, Crosshair Color, Component Based Logic
Decided to overhaul several chunks of logic into a component based system.
Health and Hitboxes are now component based.
The basic sentry AI is now also component based as an experiment.

Player has unique class name with a cheeky placeholder
Pickup Base has a clase name, not sure if useful
Experimenting with a health bar
Crosshair now colors red if your weapon is in range to attack

New friendly NPC group that is not used yet