Improved weapon data loading and structure of code
Refactored, standardized, and simplified how weapon data is loaded.

Function inside the resource file loads all the weapon data into arrays and sends it off to the weapon manager which loads it and sets it for the weapon itself, which accesses them directly.

All fields commented with descriptions for the inspector.

Names and positions of the values all standardized and better organized.

Next step is enhancing the tool logic so certain options are hidden when invalid, like charge properties if using a non-charge weapon.

Still need to implement various logic like burst fire and charge attacks.

Very pleased.